+++ src/main/java/com/example/interceptor/LoggerInterceptor.java
@@ -1,5 +1,7 @@
 package com.example.interceptor;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 
 import javax.servlet.http.HttpServletRequest;
@@ -7,11 +9,14 @@ import javax.servlet.http.HttpServletResponse;
 
 public class LoggerInterceptor extends HandlerInterceptorAdapter {
 
+    private static final Logger logger
+            = LoggerFactory.getLogger(LoggerInterceptor.class);
+
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response,
                              Object handler) {
 
-        System.out.println("[*][*] Request received from :: " + request.getRemoteAddr());
+        logger.info("[*][*] Request received from :: " + request.getRemoteAddr());
         return true;
     }
 }