+++ src/main/java/com/example/restservice/GreetingController.java
@@ -1,5 +1,7 @@
 package com.example.restservice;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,13 +11,16 @@ import java.util.concurrent.atomic.AtomicLong;
 @RestController
 public class GreetingController {
 
+    private static final Logger logger
+            = LoggerFactory.getLogger(GreetingController.class);
+    
     private static final String template = "Hello, %s!";
     private AtomicLong counter = new AtomicLong();
 
     @GetMapping("/greeting")
     public Greeting greeting(@RequestParam(value="name", defaultValue = "World") String name){
         Greeting greeting = new Greeting(counter.incrementAndGet(), String.format(template, name));
-        System.out.println("Greeting Instance : " + greeting);
+        logger.info("Greeting Instance : " + greeting);
         return greeting;
     }
 }